[chore] move utils to vite package
