+++ packages/kit/src/vite/utils.js
@@ -1,3 +1,7 @@
+import fs from 'fs';
+import path from 'path';
+import { get_runtime_path } from '../core/utils';
+
 /**
  * @param {...import('vite').UserConfig} configs
  * @returns {import('vite').UserConfig}
@@ -68,4 +72,52 @@ function merge_into(a, b) {
 			a[prop] = b[prop];
 		}
 	}
+}
+
+/** @param {import('types').ValidatedKitConfig} config */
+export function get_aliases(config) {
+	/** @type {Record<string, string>} */
+	const alias = {
+		__GENERATED__: path.posix.join(config.outDir, 'generated'),
+		$app: `${get_runtime_path(config)}/app`,
+
+		// For now, we handle `$lib` specially here rather than make it a default value for
+		// `config.kit.alias` since it has special meaning for packaging, etc.
+		$lib: config.files.lib
+	};
+
+	for (const [key, value] of Object.entries(config.alias)) {
+		alias[key] = path.resolve(value);
+	}
+
+	return alias;
+}
+
+/**
+ * Given an entry point like [cwd]/src/hooks, returns a filename like [cwd]/src/hooks.js or [cwd]/src/hooks/index.js
+ * @param {string} entry
+ * @returns {string|null}
+ */
+ export function resolve_entry(entry) {
+	if (fs.existsSync(entry)) {
+		const stats = fs.statSync(entry);
+		if (stats.isDirectory()) {
+			return resolve_entry(path.join(entry, 'index'));
+		}
+
+		return entry;
+	} else {
+		const dir = path.dirname(entry);
+
+		if (fs.existsSync(dir)) {
+			const base = path.basename(entry);
+			const files = fs.readdirSync(dir);
+
+			const found = files.find((file) => file.replace(/\.[^.]+$/, '') === base);
+
+			if (found) return path.join(dir, found);
+		}
+	}
+
+	return null;
 }
